Refactor: [session_manager] 将 SessionManagerBase 从 dingtalk 转移到 session_manager 模块
